Use base class to get distraction optimized variable
